Move client-specific stuff to extra class
